requirements.txt
python-dotenv==1.0.0
sqlalchemy==2.0.19
psycopg==3.1.8
aiogram==3.0.0rc1